Task5

+++ Tyuiu.BerezovskayaGO.Sprint3.Task5.V4.Test/DataServiceTest.cs
@@ -9,7 +9,7 @@ namespace Tyuiu.BerezovskayaGO.Sprint3.Task5.V4.Test
         {
             DataService ds = new DataService();
 
-            double x = 5.0;
+            int x = 5;
             int startValue1 = 1;
             int stopValue1 = 3;
             int startValue2 = 1;
